Reuse the common methods in serveral RandomData classes.
